requirements.txt
asgiref==3.8.1
attrs==24.2.0
cffi==1.17.1
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
django-appconf==1.0.6
django-cors-headers==4.4.0
django-storages==1.14.4
django-video-encoding==1.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
model-bakery==1.19.5
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pip==24.0
PyJWT==2.9.0
psycopg==3.2.2
psycopg-binary==3.2.2
pycparser==2.22
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
